Split label round-trip tracking into three independent specs
Each documented case (async, sync, failed delivery) now lives in its own
integration spec so failures are isolated and the scenarios can run
independently, replacing the single combined spec.
