+++ test/integrations/label_round_trip_sync/Gemfile
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+gem "waterdrop", path: ENV.fetch("WATERDROP_GEM_DIR", nil)
